backend/requirements.txt
Django
djangorestframework
celery
redis
django-cors-headers
django-redis
axios
djangorestframework-simplejwt
djangorestframework_simplejwt